+++ src/@locale/components/index.js
@@ -9,6 +9,7 @@ import map        from './map'
 import modal      from './modal'
 import pagination from './pagination'
 import picker     from './picker'
+import retry      from './retry'
 import scheduler  from './scheduler'
 import sortable   from './sortable'
 import splitPanel from './splitPanel'
@@ -26,6 +27,7 @@ const components =
     modal ,
     pagination ,
     picker ,
+    retry ,
     scheduler ,
     sortable ,
     splitPanel ,